make python-creole an optional dependency
fix #11
